modules/www/lambda_edge_authorizer/requirements.txt
boto3==1.34.48
PyJWT==2.10.1
cryptography